refactor: migrate to JitPack-compatible package structure and add installation docs
