+++ pom.xml
@@ -4,7 +4,7 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.functest.jpwise</groupId>
+    <groupId>io.github.mikeddavydov</groupId>
     <artifactId>jpwise</artifactId>
     <version>1.0-SNAPSHOT</version>
