--- src/main/java/com/functest/jpwise/core/CombinationTable.java
@@ -1,200 +1,0 @@
-/**
- * Copyright (c) 2010 Ng Pan Wei, 2013 Mikhail Davydov
- *
- * <p>Permission is hereby granted, free of charge, to any person obtaining a copy of this software
- * and associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * <p>The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * <p>THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
- * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-package com.functest.jpwise.core;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- * A collection of test case combinations generated by a test generation algorithm. This class
- * stores and manages the complete set of test cases, providing methods to access them in different
- * formats suitable for test execution.
- *
- * <p>The table can be accessed in several ways:
- *
- * <ul>
- *   <li>As a list of {@link Combination} objects
- *   <li>As a list of maps, where each map represents a test case with parameter names as keys
- *   <li>As a TestNG data provider array
- * </ul>
- *
- * <p>Example usage:
- *
- * <pre>
- * // Generate test cases
- * TestGenerator generator = new TestGenerator(input);
- * generator.generate(new PairwiseAlgorithm());
- * CombinationTable table = generator.result();
- *
- * // Access as combinations
- * for (Combination combination : table.combinations()) {
- *     log.info("Generated combination: {}", combination);
- * }
- *
- * // Access as row maps
- * List<Map<String, Object>> rows = table.asRowMapList();
- * for (Map<String, Object> row : rows) {
- *     String browser = (String) row.get("browser");
- *     String os = (String) row.get("os");
- * }
- *
- * // Use with TestNG
- * {@literal @}Test(dataProvider = "testCases")
- * public void testCombination(String description, String browser, String os) {
- *     // Test implementation
- * }
- *
- * {@literal @}DataProvider(name = "testCases")
- * public Object[][] getTestCases() {
- *     return table.asDataProvider();
- * }
- * </pre>
- */
-public class CombinationTable {
-  /** The list of test case combinations. */
-  private List<Combination> _combinations;
-
-  /** Creates a new empty combination table. */
-  public CombinationTable() {
-    super();
-    _combinations = new ArrayList<>();
-  }
-
-  /** Copy constructor. */
-  public CombinationTable(CombinationTable other) {
-    super();
-    _combinations = new ArrayList<>(other._combinations);
-  }
-
-  /**
-   * Adds a test case combination to the table.
-   *
-   * @param combination The combination to add
-   */
-  public void add(Combination combination) {
-    _combinations.add(combination);
-  }
-
-  /**
-   * Gets the number of test cases in the table.
-   *
-   * @return The number of combinations
-   */
-  public int size() {
-    return _combinations.size();
-  }
-
-  /**
-   * Gets all test case combinations.
-   *
-   * @return The list of combinations
-   */
-  public List<Combination> combinations() {
-    return new ArrayList<>(_combinations);
-  }
-
-  /**
-   * Converts the test cases to a list of maps. Each map represents a test case where keys are
-   * parameter names and values are the parameter values. A special key "combination_description"
-   * contains the string representation of the combination.
-   *
-   * @return A list of maps representing the test cases
-   */
-  public List<Map<String, Object>> asRowMapList() {
-    List<Map<String, Object>> rows = new ArrayList<>();
-
-    for (Combination combination : _combinations) {
-      Map<String, Object> row = new HashMap<>();
-
-      for (EquivalencePartition value : combination.getValues()) {
-        row.put(value.getParentParameter().getName(), value.getValue());
-      }
-      row.put("combination_description", combination.toString());
-      rows.add(row);
-    }
-
-    return rows;
-  }
-
-  /**
-   * Converts the test cases to a format suitable for TestNG's data provider. Each row in the
-   * returned array contains: - The combination description as the first element - The actual
-   * parameter values in subsequent elements
-   *
-   * @return A 2D array suitable for use with TestNG's data provider
-   */
-  public Object[][] asDataProvider() {
-    List<Object[]> testCases = new ArrayList<>();
-    for (Combination combination : _combinations) {
-      testCases.add(combination.asDataProviderRow());
-    }
-
-    Object[][] res = new Object[testCases.size()][];
-    return testCases.toArray(res);
-  }
-
-  /**
-   * Gets the number of parameters in each combination. Returns -1 if the table is empty.
-   *
-   * @return The number of parameters, or -1 if empty
-   */
-  public int breadth() {
-    if (_combinations.isEmpty()) {
-      return -1;
-    }
-    Combination combination = _combinations.get(0);
-    return combination.size();
-  }
-
-  /**
-   * Calculates the total number of unique parameter value pairs covered by the test cases in this
-   * table. This is useful for evaluating the coverage of pairwise testing.
-   *
-   * @return The number of unique parameter value pairs covered
-   */
-  public int span() {
-    int depth = breadth();
-    if (depth < 0) return -1;
-    HashMap<String, Integer> tempMap = new HashMap<>();
-    for (int i = 0; i < depth; i++) {
-      for (int j = i + 1; j < depth; j++) {
-        for (Combination generatedEntry : _combinations) {
-          if ((generatedEntry.getValue(i) != null) && (generatedEntry.getValue(j) != null)) {
-            Combination entry = new Combination(depth);
-            entry.setValue(i, generatedEntry.getValue(i));
-            entry.setValue(j, generatedEntry.getValue(j));
-            String key = entry.getKey();
-            tempMap.put(key, 1);
-          }
-        }
-      }
-    }
-    return tempMap.size();
-  }
-
-  @Override
-  public String toString() {
-    return String.format(
-        "CombinationTable{%d combinations.%s}",
-        _combinations.size(),
-        _combinations.isEmpty() ? "" : (" First is: " + _combinations.get(0)));
-  }
-}
