--- src/main/java/com/functest/jpwise/core/EquivalencePartition.java
@@ -1,16 +1,0 @@
-package com.functest.jpwise.core;
-
-/**
- * @author DavydovMD Date: 17.06.13 Time: 15:49
- */
-public interface EquivalencePartition {
-  TestParameter getParentParameter();
-
-  void setParentParameter(TestParameter parameter);
-
-  Object getValue();
-
-  String getName();
-
-  boolean isCompatibleWith(EquivalencePartition v2);
-}
