--- src/test/java/com/functest/jpwise/JpWiseDataProviderDemoTest.java
@@ -1,99 +1,0 @@
-package com.functest.jpwise;
-
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertNotNull;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
-
-import com.functest.jpwise.algo.PairwiseAlgorithm;
-import com.functest.jpwise.core.CompatibilityPredicate;
-import com.functest.jpwise.core.EquivalencePartition;
-import com.functest.jpwise.core.SimpleValue;
-import com.functest.jpwise.core.TestGenerator;
-import com.functest.jpwise.core.TestInput;
-import com.functest.jpwise.core.TestParameter;
-
-/**
- * Demonstrates how to use JPWise with TestNG's data provider feature. This allows running the same
- * test with different parameter combinations generated by the pairwise algorithm.
- */
-public class JpWiseDataProviderDemoTest {
-  private static final Logger log = LoggerFactory.getLogger(JpWiseDataProviderDemoTest.class);
-
-  @DataProvider(name = "browserConfigurations")
-  public Object[][] getBrowserConfigurations() {
-    // Define test parameters
-    TestParameter browser =
-        new TestParameter(
-            "browser",
-            Arrays.<EquivalencePartition>asList(
-                SimpleValue.of("Chrome", "116.0"),
-                SimpleValue.of("Firefox", "118.0"),
-                SimpleValue.of("Safari", "17.0")));
-
-    TestParameter os =
-        new TestParameter(
-            "os",
-            Arrays.<EquivalencePartition>asList(
-                SimpleValue.of("Windows", "10.0"),
-                SimpleValue.of("macOS", "14.1"),
-                SimpleValue.of("Linux", "6.5")));
-
-    // Define compatibility rules (Safari only works with macOS)
-    List<CompatibilityPredicate> rules =
-        Arrays.asList(
-            (EquivalencePartition v1, EquivalencePartition v2) -> {
-              // Only apply rules if we're dealing with browser and OS parameters
-              if (!(v1.getParentParameter().getName().equals("browser")
-                  && v2.getParentParameter().getName().equals("os"))) {
-                return true;
-              }
-
-              // Safari only works with macOS
-              if (v1.getName().equals("Safari")) {
-                return v2.getName().equals("macOS");
-              }
-              return true;
-            });
-
-    // Create test input
-    TestInput input = new TestInput();
-    input.add(new TestParameter("browser", browser.getPartitions(), rules));
-    input.add(new TestParameter("os", os.getPartitions()));
-
-    // Generate combinations
-    TestGenerator generator = new TestGenerator(input);
-    generator.generate(new PairwiseAlgorithm());
-
-    // Return as data provider array
-    return generator.result().asDataProvider();
-  }
-
-  @Test(dataProvider = "browserConfigurations")
-  public void testBrowserCompatibility(String description, String browser, String os) {
-    // description is automatically generated by JPWise and includes the combination details
-    assertNotNull(description, "Description should not be null");
-    assertNotNull(browser, "Browser should not be null");
-    assertNotNull(os, "OS should not be null");
-
-    // Verify Safari-macOS compatibility rule
-    if (browser.equals("Safari")) {
-      assertEquals(os, "macOS", "Safari should only run on macOS");
-    }
-
-    // Simulate browser testing
-    log.info("JpWiseDataProviderDemoTest: Testing browser configuration: {} on {}", browser, os);
-
-    // Your actual test logic would go here
-    // For example:
-    // - Launch browser
-    // - Run tests
-    // - Verify results
-  }
-}
